Added partial modification and simplified modified the pipe(). Modified the rust user programs to adjust to the new api.

+++ user/src/lib.rs
@@ -78,7 +78,7 @@ pub fn open(path: &str, flags: OpenFlags) -> isize {
 pub fn close(fd: usize) -> isize {
     sys_close(fd)
 }
-pub fn pipe(pipe_fd: &mut [usize]) -> isize {
+pub fn pipe(pipe_fd: &mut [i32]) -> isize {
     sys_pipe(pipe_fd)
 }
 pub fn read(fd: usize, buf: &mut [u8]) -> isize {
